изменен тип данных

+++ src/main/java/dataBaseHandler/ResultEntity.java
@@ -1,7 +1,6 @@
 package dataBaseHandler;
 
 import jakarta.persistence.*;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -15,13 +14,13 @@ public class ResultEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Integer x;
+    private Double x;
     private Double y;
     private Float R;
     private Date date;
     private boolean status;
 
-    public ResultEntity(Integer x, Double y, Float r, Date date, boolean status) {
+    public ResultEntity(Double x, Double y, Float r, Date date, boolean status) {
         this.x = x;
         this.y = y;
         this.R = r;
